+++ src/main/java/ongi/ongibe/domain/album/entity/Picture.java
@@ -1,4 +1,4 @@
-package ongi.ongibe.domain;
+package ongi.ongibe.domain.album.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -14,7 +14,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import ongi.ongibe.domain.album.entity.Album;
 import ongi.ongibe.domain.user.entity.User;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.SQLDelete;
