+++ src/main/java/ongi/ongibe/domain/album/entity/Place.java
@@ -1,4 +1,4 @@
-package ongi.ongibe.domain;
+package ongi.ongibe.domain.album.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
